Clean up the UI imports into components
